Parameterize render sanitization + Deep general sanitization
