+++ tests/helpers.test.js
@@ -500,7 +500,7 @@ describe('Templates', () => {
     test('inserts template with different positions', () => {
       container.innerHTML = '<p>original</p>'
 
-      insertTemplate('.test-container', 'template1', 'end content', 'end')
+      insertTemplate('.test-container', 'template1', 'end content', { position: 'end' })
       expect(container.innerHTML).toBe('<p>original</p><div>end content</div>')
     })
 
@@ -524,6 +524,90 @@ describe('Templates', () => {
       expect(() => {
         insertTemplate('.test-container', 'nonExistentTemplate', 'test')
       }).toThrow("[Ralix] Template 'nonExistentTemplate' not found")
+    })
+  })
+
+  describe('render', () => {
+    describe('XSS Protection', () => {
+      test('should sanitize javascript: URLs in attributes', () => {
+        const maliciousData = {
+          src: 'javascript:alert("XSS")',
+          alt: 'Test image'
+        }
+
+        const result = render('attributeInjection', maliciousData)
+
+        expect(result).not.toContain('javascript:')
+        expect(result).not.toContain('alert("XSS")')
+        expect(result).toContain('alt="Test image"')
+      })
+
+      test('should preserve template structure while sanitizing only variables', () => {
+        const data = {
+          userContent: '<script>alert("XSS")</script>Safe content',
+          userImage: 'javascript:alert("XSS")'
+        }
+
+        const result = render('structureTemplate', data)
+
+        // Template structure should be preserved
+        expect(result).toContain('<script type="application/json">')
+        expect(result).toContain('onclick="handleClick()"')
+        expect(result).toContain('onerror="fallback()"')
+
+        // But user variables should be sanitized
+        expect(result).not.toContain('alert("XSS")')
+        expect(result).toContain('Safe content')
+        expect(result).toContain('<img src="" onerror="fallback()">')
+      })
+
+      test('should sanitize nested object variables', () => {
+        const data = {
+          user: {
+            name: '<script>alert("name")</script>John',
+            profile: {
+              bio: '<img onerror="alert(1)" src="x">Developer'
+            }
+          },
+          items: [
+            '<script>alert("item1")</script>Item 1',
+            'javascript:alert("item2")',
+            'Safe item'
+          ]
+        }
+
+        const result = render('nestedTemplate', data)
+
+        // Scripts should be removed from variables
+        expect(result).not.toContain('<script>')
+        expect(result).not.toContain('alert(')
+        expect(result).not.toContain('javascript:')
+
+        // Safe content should be preserved
+        expect(result).toContain('John')
+        expect(result).toContain('Developer')
+        expect(result).toContain('Safe item')
+        expect(result).toContain('<img src="x">')
+      })
+
+      test('should handle primitive values in data', () => {
+        const data = {
+          message: '<script>alert("XSS")</script>Hello',
+          count: 42,
+          isActive: true,
+          nullValue: null,
+          undefinedValue: undefined
+        }
+
+        const result = render('primitiveTemplate', data)
+
+        expect(result).not.toContain('<script>')
+        expect(result).toContain('Hello')
+        expect(result).toContain('Count: 42')
+        expect(result).toContain('Active: true')
+        expect(result).toContain('Null: ')
+        expect(result).toContain('Undefined: ')
+      })
     })
   })
 })
